Fixed skill completion time and skill completion notification.  EvilNumber class beginning to be expanded and corrected.  Notable correction is with the LOGIC_OPERATOR code generation macro where the 'a' parameter was not even being used to substitute in the actual operator, instead, the '<' operator was hard-coded in the macro.

+++ src/eve-common/utils/EvilNumber.cpp
@@ -16,7 +16,7 @@ PyRep* EvilNumber::GetPyObject()
         if ( mValue.iVal > INT_MAX || mValue.iVal < INT_MIN)
             return (PyRep*)new PyLong(mValue.iVal);
         else
-            return (PyRep*)new PyInt(mValue.iVal);
+            return (PyRep*)new PyInt((int32)(mValue.iVal));
     }
     else if (mType == evil_number_float)
         return (PyRep*)new PyFloat(mValue.fVal);
